refactor: add event propagation for click & key events; try to use const (replace into separate files)
