+++ src/css/style.css
@@ -86,4 +86,3 @@
 .arrow-right::before {
   content: '\25B6';
 }
-
