+++ src/js/languages.js
@@ -1,0 +1,24 @@
+const engLang = [
+  '`', '1', '2', '3', '4', '5', '6', '7', '8', '9', '0', '-', '=', 'Backspace',
+  'Tab', 'q', 'w', 'e', 'r', 't', 'y', 'u', 'i', 'o', 'p', '[', ']', 'Delete',
+  'CapsLock', 'a', 's', 'd', 'f', 'g', 'h', 'j', 'k', 'l', ';', '"', 'Enter',
+  'Shift', 'z', 'x', 'c', 'v', 'b', 'n', 'm', ',', '.', '/', 'up',
+  'Control', 'language', 'Alt', 'space', 'left', 'down', 'right',
+];
+
+const ruLang = [
+  'ё', '1', '2', '3', '4', '5', '6', '7', '8', '9', '0', '-', '=', 'Backspace',
+  'Tab', 'й', 'ц', 'у', 'к', 'е', 'н', 'г', 'ш', 'щ', 'з', 'х', 'ъ', 'Delete',
+  'CapsLock', 'ф', 'ы', 'в', 'а', 'п', 'р', 'о', 'л', 'д', 'ж', 'э', 'Enter',
+  'Shift', 'я', 'ч', 'с', 'м', 'и', 'т', 'ь', 'ь', 'б', 'ю', 'up',
+  'Control', 'language', 'Alt', 'space', 'left', 'down', 'right',
+];
+
+const languages = {
+  en: engLang,
+  ru: ruLang,
+};
+
+const insertLineBreak = (key) => ['Backspace', 'Delete', 'Enter', 'up'].includes(key);
+
+export { languages, insertLineBreak };
